Make Shop into an ISH
This allows for the Shop to be passed as ingredientsHostLocator to a Producer. The effect of that is that the Producer request the ingredients from the Shop. The Shop than (by relaying to the OrderItems service) tries retrieving the ingredients from the storage. If  that fails it tries to procure the ingredients again with another call to a Producer.

This recursive order call should result in a successfull items order if the (sub, and sub-sub, etc) ingredients are present in the storage and/ or can be produced by the Producer s.