fix: building Silo causes unexpected hole
Interesting issue uncovered by recent change of MoveTo enabling ground level movement:

After the coremove change building the Silo caused a strange extra hole at the ground when building the first storage Chest layer.

The cause for the hole was the following:
- when the first storage layer needs to be build the BuildBlueprint_Task in tole_builder attempts to GoTo the start of the layer startPoint.
- this call however fails because there are blocks in the way to get there (either some normal block underground or the chests of a previous Silo that was discarded), and the GoTo call was called with force = false
- hence, with the newly added MoveTo functionality, the turtle moves above ground (z=2), i.e. out of the shaft, and tries again.
- when it than moves down it creates the hole.

This error never really showed before because
- before the new MoveTo functionality, the MoveTo would fails after a few attempts.
- than role_builder would ignore the failure of GoTo, and attempt to build the layer anyway
- it would do so with a coremove.MoveTo statement that DOES force the move, and hence it would get to the start point anyway.

So, the mistake in the logic happened before, but we didn't really see it, because it was masked by the behaviour of role_bulder.

We now fixed this issue with this commit,

Hope it works everywhere.... !

by an existing issue in the

that was hidden by the fact that role_build