manufacturing: revise getFuelNeed_Production_Att + how/ where it is used
getFuelNeed_Production_Att is now defined as the fuelNeed for producing items assuming the ingredients (incl possible production fuel) are available (in a Turtle located) at the Factory baseLocation and the results are to be delivered to that Location. In other worths we ignore fuel needs to and from the Factory.

ProductionSpot:
- add and test getFuelNeed_Production_Att

Factory:
- change and test getFuelNeed_Production_Att to new definition
- replace NeedsTo_ProvideItemsTo_SSrv by a call to getFuelNeed_Production_Att (and remove NeedsTo_ProvideItemsTo_SSrv )
- remove getProductionLocation_Att

enterprise_energy:
- use base factoryLocation to calculate GetFuelNeed_Refuel_Att (instead of getProductionLocation_Att)